Remove styledButton.tsx
Move code over to button.tsx
