+++ src/index.tsx
@@ -12,7 +12,6 @@ export * from "./header";
 export * from "./spinner";
 export * from "./util";
 export * from "./modal";
-export * from "./styledButton";
 
 
 /**
